+++ CHANGELOG.md
@@ -8,6 +8,23 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+
+## v2.0.0 - 2017-10-31
+### Changed
+- ソースコードのディレクトリ構造を変更。
+- sskts-domain@v23(スキーマ一新)に対応。
+
+### Security
+- install latest packages.
+
 ## v1.1.1 - 2017-05-18
 ### Added
 - blueprintからapiドキュメントを生成してdocsに追加。